[test] Extract `for-injection`
