+++ tests/src/when-component-associative-header.cpp
@@ -8,32 +8,7 @@
 
 #include "pbdt/bdd.hpp"
 
-namespace pbdt::bdd::detail
-{
-    template <typename... Domains, typename Domain>
-    constexpr auto operator+(WhenContext<Domains...>&& context, Domain&& domain)
-    {
-        return context.andWhen("", std::forward<Domain>(domain));
-    }
-}
-
-struct ToContainer
-{
-    template <typename T>
-    constexpr auto operator()(T&& t) const
-    {
-        return exstd::toContainer(std::forward<T>(t));
-    }
-};
-
-struct When
-{
-    template <typename Domain>
-    constexpr auto operator()(Domain&& domain) const
-    {
-        return pbdt::bdd::when("", std::forward<Domain>(domain));
-    }
-};
+#include "for-injection.hpp"
 
 int main()
 {
